Improved: moved services from Util to User
