+++ src/services/UtilService.ts
@@ -9,14 +9,6 @@ const fetchRoles = async (payload: any): Promise<any> => {
   })
 }
 
-const getUserAssociatedProductStores = async (payload: any): Promise<any> => {
-  return api({
-    url: 'performFind',
-    method: 'POST',
-    data: payload
-  })
-}
-
 const getSecurityGroups = async (payload: any): Promise<any> => {
   return api({
     url: "performFind",
@@ -26,23 +18,7 @@ const getSecurityGroups = async (payload: any): Promise<any> => {
   })
 }
 
-const createProductStoreRole = async (payload: any): Promise <any> => {
-  return api({
-    url: "service/createProductStoreRole", 
-    method: "post",
-    data: payload
-  });
-}
-
-const updateProductStoreRole = async (payload: any): Promise <any> => {
-  return api({
-    url: "service/updateProductStoreRole", 
-    method: "post",
-    data: payload
-  });
-}
-
-const fetchFacilities = async(query: any): Promise<any> => {
+const fetchFacilities = async (query: any): Promise<any> => {
   return api({
     url: "performFind",
     method: "get",
@@ -51,63 +27,7 @@ const fetchFacilities = async(query: any): Promise<any> => {
   })
 }
 
-const getUserSecurityGroup = async (payload: any): Promise<any> => {
-  return api({
-    url: "performFind",
-    method: "POST",
-    data: payload
-  })
-}
-
-const updateUserSecurityGroup = async (payload: any): Promise <any> => {
-  return api({
-    url: "service/updateUserLoginToSecurityGroup", 
-    method: "post",
-    data: payload
-  });
-}
-
-const addUserToSecurityGroup = async (payload: any): Promise <any> => {
-  return api({
-    url: "service/addUserLoginToSecurityGroup", 
-    method: "post",
-    data: payload
-  });
-}
-
-const createPartyRole = async (payload: any): Promise <any> => {
-  return api({
-    url: "service/ensurePartyRole", 
-    method: "post",
-    data: payload
-  });
-}
-
-const getUserAssociatedFacilities = async (payload: any): Promise<any> => {
-  return api({
-    url: "performFind",
-    method: "POST",
-    data: payload
-  })
-}
-
-const addPartyToFacility = async (payload: any): Promise <any> => {
-  return api({
-    url: "service/addPartyToFacility", 
-    method: "post",
-    data: payload
-  });
-}
-
-const removePartyFromFacility = async (payload: any): Promise <any> => {
-  return api({
-    url: "service/removePartyFromFacility", 
-    method: "post",
-    data: payload
-  });
-}
-
-const fetchProductStores = async(query: any): Promise<any> => {
+const fetchProductStores = async (query: any): Promise<any> => {
   return api({
     url: "performFind",
     method: "get",
@@ -117,19 +37,9 @@ const fetchProductStores = async(query: any): Promise<any> => {
 }
 
 export const UtilService = {
-  addPartyToFacility,
-  createProductStoreRole,
-  addUserToSecurityGroup,
   fetchFacilities,
   fetchProductStores,
   getSecurityGroups,
-  fetchRoles,
-  getUserAssociatedProductStores,
-  getUserAssociatedFacilities,
-  getUserSecurityGroup,
-  removePartyFromFacility,
-  updateProductStoreRole,
-  updateUserSecurityGroup,
-  createPartyRole,
+  fetchRoles
 }
-  
+
